move ignoreAll function to its own file
